+++ spring-batch-example-master/src/main/java/com/eai/BatchJobCDL/DTO/EscDTO.java
@@ -1,5 +1,7 @@
 package com.eai.BatchJobCDL.DTO;
 
+import java.math.BigDecimal;
+
 public class EscDTO {
 	
 	  private String age;
@@ -7,7 +9,7 @@ public class EscDTO {
 	  private String type;
 	  private String cpt;
 	  private String raisonSocial;
-	  private Double montantCreance;
+	  private BigDecimal montantCreance;
 	  private String dateCreance;
 	  private String noDossier;
 	  private String refferenceValeur;
@@ -17,7 +19,7 @@ public class EscDTO {
 	  private String dateRemise;
 	  private String referenceLiaison;
 	  private String codeRejet;
-	  private Double commission;
+	  private BigDecimal commission;
 	  private String field17;
 	  private String field18;
 	public String getAge() {
@@ -50,10 +52,10 @@ public class EscDTO {
 	public void setRaisonSocial(String raisonSocial) {
 		this.raisonSocial = raisonSocial;
 	}
-	public Double getMontantCreance() {
+	public BigDecimal getMontantCreance() {
 		return montantCreance;
 	}
-	public void setMontantCreance(Double montantCreance) {
+	public void setMontantCreance(BigDecimal montantCreance) {
 		this.montantCreance = montantCreance;
 	}
 	public String getDateCreance() {
@@ -105,10 +107,10 @@ public class EscDTO {
 	public void setCodeRejet(String codeRejet) {
 		this.codeRejet = codeRejet;
 	}
-	public Double getCommission() {
+	public BigDecimal getCommission() {
 		return commission;
 	}
-	public void setCommission(Double commission) {
+	public void setCommission(BigDecimal commission) {
 		this.commission = commission;
 	}
 	public String getField18() {
